requirements.txt
cython==0.29.28
flask==2.0.1
requests==2.26.0
scikit-learn==0.24.2
fast-histogram==0.10
lz4==3.1.3
pycocotools==2.0.4
orjson==3.6.4
pandas==1.3.4
python-dotenv==0.19.2
pymongo==3.12.1
scipy==1.8.0
boto3==1.21.2
google-cloud-storage===2.1.0
isodate==0.6.1
pyuwsgi==2.0.21
gevent==21.12.0
s3urls==0.0.3
simplejson==3.17.6
redis==4.3.4
psycopg2==2.9.3
SQLAlchemy==1.4.39
numpy # Leave this unpinned so pip can accomodate version conflicts.
flatdict==4.0.1
sqlglot==6.2.7
evaluate==0.2.2
seqeval==1.2.2
sentry-sdk[flask]==1.15.0
psutil==5.9.4
opencv-python-headless==4.6.0.66

hdbscan==0.8.28
joblib==1.1.0
umap-learn==0.5.3
